requirements.txt
pandas==1.5.3
scikit-learn==1.3.0
streamlit==1.28.0
joblib==1.3.2
python-whois==0.8.0
urlextract==1.8.0
scipy==1.11.1
numpy==1.24.3
plotly==5.17.0